+++ src/WordleGame/leveled/GamePlay.java
@@ -1,4 +1,4 @@
-package WordleGame.leveling;
+package WordleGame.leveled;
 
 // 게임 플레이를 실행하는 추상메서드
 // 추상메서드로 게임 레벨별로 분기가 가능하게 설정
@@ -8,7 +8,7 @@ abstract class GamePlay extends RandomList {
     protected int tryCount2 = 0;
     protected int tryCount3 = 0;
 
-    public abstract void Gameresult();
+    public abstract void gamePlay();
 
-    public abstract void Gameplay();
+    public abstract void gameResult();
 }
